03K: Hang during reduce pami test (D173943)
* A bug in PAMI_DOUBLE, PAMI_BAND flow was found and fixed

+++ algorithms/protocols/allreduce/ReduceFunctions.h
@@ -283,7 +283,7 @@ namespace CCMI
                     func = (coremath) Core_int32_band;
                     break;
                   case PAMI_DOUBLE:
-                    sizeOfType = sizeof(float);
+                    sizeOfType = sizeof(double);
                     func = (coremath) Core_uint64_band;
                     break;
                   case PAMI_SIGNED_LONG_LONG:
